perf(frontend): 已下架主机单据展示调整 #10751 # Reviewed, transaction id: 44198
