+++ dbm-ui/frontend/src/views/ticket-center/common/ticket-detail/components/task-info/com-factory/resource-pool-strategy/RecycleOldHost.vue
@@ -24,10 +24,81 @@
         {{ ticketDetails.details.parent_ticket }}
       </BkButton>
     </InfoItem>
+    <InfoItem :label="t('已下架主机')">
+      <BkTable :data="ticketDetails.details.recycle_hosts">
+        <BkTableColumn
+          field="ip"
+          fixed="left"
+          label="IP"
+          :min-width="150">
+          <template #header>
+            <div class="ip-header">
+              IP
+              <DbIcon
+                type="copy"
+                @click="copyAllIp" />
+            </div>
+          </template>
+        </BkTableColumn>
+        <BkTableColumn
+          field="bk_cloud_name"
+          :label="t('管控区域')"
+          :min-width="120" />
+        <BkTableColumn
+          field="status"
+          :label="t('Agent 状态')"
+          :min-width="120">
+          <template #default="{ data }: { data: RowData }">
+            <HostAgentStatus :data="data.status" />
+          </template>
+        </BkTableColumn>
+        <BkTableColumn
+          field="city"
+          :label="t('地域')"
+          :min-width="120">
+          <template #default="{ data }: { data: RowData }">
+            {{ data.city || '--' }}
+          </template>
+        </BkTableColumn>
+        <BkTableColumn
+          field="sub_zone"
+          :label="t('园区')"
+          :min-width="120">
+          <template #default="{ data }: { data: RowData }">
+            {{ data.sub_zone || '--' }}
+          </template>
+        </BkTableColumn>
+        <BkTableColumn
+          field="rack_id"
+          :label="t('机架')"
+          :min-width="120">
+          <template #default="{ data }: { data: RowData }">
+            {{ data.rack_id || '--' }}
+          </template>
+        </BkTableColumn>
+        <BkTableColumn
+          field="bk_os_name"
+          :label="t('操作系统')"
+          :min-width="120">
+          <template #default="{ data }: { data: RowData }">
+            {{ data.bk_os_name || '--' }}
+          </template>
+        </BkTableColumn>
+        <BkTableColumn
+          field="device_class"
+          :label="t('机型')"
+          :min-width="120">
+          <template #default="{ data }: { data: RowData }">
+            {{ data.device_class || '--' }}
+          </template>
+        </BkTableColumn>
+        <BkTableColumn
+          field="remark"
+          :label="t('备注')"
+          :min-width="200" />
+      </BkTable>
+    </InfoItem>
   </InfoList>
-  <RecycleHostCard
-    :data="ticketDetails.details.recycle_hosts"
-    :title="t('已下架主机')" />
 </template>
 <script setup lang="ts">
   import { useI18n } from 'vue-i18n';
@@ -37,15 +108,19 @@
 
   import { TicketTypes } from '@common/const';
 
-  import { getBusinessHref } from '@utils';
+  import HostAgentStatus from '@components/host-agent-status/Index.vue';
+
+  import { execCopy, getBusinessHref } from '@utils';
 
   import InfoList, { Item as InfoItem } from '../components/info-list/Index.vue';
-
-  import RecycleHostCard from './RecycleHostCard.vue';
 
   interface Props {
     ticketDetails: TicketModel<Common.ResourcePoolRecycle>;
   }
+
+  type RowData = Props['ticketDetails']['details']['recycle_hosts'][number];
+
+  type Emits = (e: 'hide-remark') => void;
 
   defineOptions({
     name: TicketTypes.RECYCLE_OLD_HOST,
@@ -53,6 +128,8 @@
   });
 
   const props = defineProps<Props>();
+
+  const emits = defineEmits<Emits>();
 
   const { t } = useI18n();
   const router = useRouter();
@@ -66,4 +143,15 @@
     });
     window.open(getBusinessHref(href, props.ticketDetails.bk_biz_id), '_blank');
   };
+
+  const copyAllIp = () => {
+    const ips = props.ticketDetails.details.recycle_hosts.map((item) => item.ip);
+    if (ips.length > 0) {
+      execCopy(ips.join('\n'), t('复制成功，共n条', { n: ips.length }));
+    }
+  };
+
+  onMounted(() => {
+    emits('hide-remark');
+  });
 </script>
